\name{print.lognlm}
\alias{print.lognlm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print method for the lognlm class
}
\description{
Printing the most important feautures of a \code{'lognlm'} model.
}
\usage{
\method{print}{lognlm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object of class \code{segmented} }
  \item{digits}{ number of digits to be printed }
  \item{\dots}{ arguments passed to other functions }
}
\author{ Vito M.R. Muggeo }
\seealso{ \code{\link{summary.lognlm}}, \code{\link{print.summary.lognlm}} }

\keyword{ models }
